refactor: reorganize ConversationEventRow components
- Rename ConversationContent to ConversationStream
- Move ConversationStream to its own folder at components/internal/ConversationStream/
- Move ConversationEventRow and its stories under ConversationStream folder
- Extract MessageContent components (UserMessageContent, AssistantMessageContent, UnknownMessageContent) into separate files under MessageContent/ subfolder
- Update all imports to reflect new structure
- Add index files for cleaner imports

This improves code organization, readability, and testability by creating a clearer separation of concerns.
