+++ humanlayer-wui/src/components/internal/ConversationStream/MessageContent/UnknownMessageContent.tsx
@@ -1,0 +1,4 @@
+
+export function UnknownMessageContent() {
+  return <div>Unknown Message Content</div>
+}
