+++ humanlayer-wui/src/components/internal/ConversationStream/index.ts
@@ -1,0 +1,3 @@
+export { ConversationStream } from './ConversationStream'
+export { ConversationEventRow } from './ConversationEventRow'
+export type { ConversationEventRowProps } from './ConversationEventRow'
